+++ log-all-the-things/common/src/main/java/com/chimericdream/logallthethings/lavalog/LavaLogProperties.java
@@ -1,0 +1,10 @@
+package com.chimericdream.logallthethings.lavalog;
+
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+
+public final class LavaLogProperties {
+    public static final BooleanProperty LAVALOGGED = BooleanProperty.create("lavalogged");
+
+    private LavaLogProperties() {
+    }
+}